mysite/pruebas/requirements_pruebas.txt
# ============================================
# DEPENDENCIAS PARA SISTEMA DE CORREOS Y REPORTES
# ============================================
# Instalar con: pip install -r requirements_pruebas.txt

# === Envío de correos ===
python-dotenv==1.0.0         # Para cargar variables de entorno desde .env

# === Generación de reportes Excel ===
openpyxl==3.1.2              # Para crear y manipular archivos Excel (.xlsx)
pandas==2.1.4                # Para manipulación de datos (opcional pero recomendado)
xlsxwriter==3.1.9            # Alternativa para generar Excel con más opciones de formato

# === APIs de almacenamiento en la nube (para futuro) ===
# google-api-python-client==2.108.0    # Para Google Drive API
# google-auth-httplib2==0.2.0          # Autenticación Google
# google-auth-oauthlib==1.2.0          # OAuth Google
# msal==1.25.0                         # Para Microsoft Graph API (OneDrive)

# === Utilidades ===
requests==2.31.0             # Para hacer peticiones HTTP
